requirements.txt
Flask==1.1.1
xgboost
pandas
sklearn
matplotlib
git+https://github.com/mitsuhiko/flask-oauth
responder
loguru
request
jinja2
gcloud
google-api-python-client
PyMySQL
flask-sqlalchemy